style: extract podmodal, helpmodal
