+++ wondrous-app/components/Common/SidebarMainPods/index.tsx
@@ -1,4 +1,4 @@
-import { PodModal } from 'components/Common/SidebarMainPods/PodModal';
+import { PodModal } from 'components/Common/PodModal';
 import { PodsButton, PodsIconWrapper } from 'components/Common/SidebarMainPods/styles';
 import SidebarTooltip from 'components/Common/SidebarMainTooltip';
 import PodsIcon from 'components/Icons/Sidebar/podsGradient.svg';
